requirements.txt
feedparser==6.0.11
PyYAML==6.0.2
requests==2.32.3
pytest==8.4.1
aiosmtpd==1.4.6
